Bugfix Magento 2.2 Set Payment Information returning 400 error.
The reason for this change is that the checkout is blocked when Agreements should be accepted.
Buckaroo actively updates the payment information in the checkout summary, which comes across this check and thus gets blocked when checkout agreements are required.

With this update the Agreements Validaiton that is called on setting the payment information (which seems to be unneccesary from Magento's side) get ignored.
This is done through the use of a preference.